Move stdlib.c to briey/lib, add UART out to memtest
