+++ projects/briey/dhrystone/makefile
@@ -11,7 +11,7 @@ LDFLAGS += -lc
 CFLAGS +=  -DPREALLOCATE=1 -DHOST_DEBUG=0 -DMSC_CLOCK 
 
 
-SRCS = 	src/main.c src/dhry_1.c src/dhry_2.c src/crt.S  src/stdlib.c 
+SRCS = 	src/main.c src/dhry_1.c src/dhry_2.c src/crt.S  ../libs/stdlib.c 
 LDFLAGS += -lc #Used to get memcpy/strcmp stuff only
 
 INC += -I../../../libs/
